+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # 2.15.0 - TBD
+- Changed the format of the "group" key in the extracted problem data so that it is consistent across parsers
 - Added a parser for mrJudge
 - Fixed the Luogu and the Codeforces parser
 - Removed the A2OJ parser